refactor: extract tests from window.rs to separate file
- Created `crates/laminar-core/src/operator/window/tests.rs`
- Moved approximately 2000 lines of tests from `window.rs`
- Updated `window.rs` to include `mod tests` as an external submodule
- Improved maintainability by reducing `window.rs` file size
- Applied formatting with `cargo fmt`
